refactor(handlers): extract manage.ts handler from intent.ts (Stage B)
handleSAPManage (feature probe, discovery cache, package CRUD, FLP catalog
mgmt) → its own module importing leaf helpers; intent.ts imports it for
dispatch. First handler extraction. 3672 tests, A1 zero-diff.
